+++ NINA.Plugin.TargetScheduler/Planning/Exposures/BasicExposureSelector.cs
@@ -18,22 +18,26 @@ namespace NINA.Plugin.TargetScheduler.Planning.Exposures {
 
         public IExposure Select(DateTime atTime, IProject project, ITarget target) {
             if (AllExposurePlansRejected(target)) {
-                throw new Exception($"unexpected: all exposure plans were rejected at exposure selection time for target '{target.Name}' at time {atTime}");
+                TSLogger.Warning($"unexpected: all exposure plans were rejected at exposure selection time for target '{target.Name}' at time {atTime}");
+                return null;
             }
 
-            foreach (IFilterCadenceItem item in FilterCadence) {
-                IExposure exposure = target.AllExposurePlans[item.ReferenceIdx];
-                if (exposure.Rejected) { continue; }
+            try {
+                foreach (IFilterCadenceItem item in FilterCadence) {
+                    IExposure exposure = target.AllExposurePlans[item.ReferenceIdx];
+                    if (exposure.Rejected) { continue; }
 
-                exposure.PreDither = DitherManager.DitherRequired(exposure);
-                FilterCadence.SetLastSelected(item);
-                return exposure;
+                    exposure.PreDither = DitherManager.DitherRequired(exposure);
+                    FilterCadence.SetLastSelected(item);
+                    return exposure;
+                }
+            } catch (Exception ex) {
+                TSLogger.Warning($"exception in basic exposure selector for target '{target.Name}' at time {atTime}: {ex.Message}, aborting target");
+                return null;
             }
 
-            // Fail safe ... should not happen
-            string msg = $"unexpected: no acceptable exposure plan in basic exposure selector for target '{target.Name}' at time {atTime}";
-            TSLogger.Error(msg);
-            throw new Exception(msg);
+            TSLogger.Warning($"no acceptable exposure plan in basic exposure selector for target '{target.Name}' at time {atTime}");
+            return null;
         }
 
         public void ExposureTaken(IExposure exposure) {