+++ NINA.Plugin.TargetScheduler/Planning/Exposures/OverrideOrderExposureSelector.cs
@@ -17,28 +17,32 @@ namespace NINA.Plugin.TargetScheduler.Planning.Exposures {
 
         public IExposure Select(DateTime atTime, IProject project, ITarget target) {
             if (AllExposurePlansRejected(target)) {
-                throw new Exception($"unexpected: all exposure plans were rejected at exposure selection time for target '{target.Name}' at time {atTime}");
+                TSLogger.Warning($"unexpected: all exposure plans were rejected at exposure selection time for target '{target.Name}' at time {atTime}");
+                return null;
             }
 
-            bool preDither = false;
-            foreach (IFilterCadenceItem item in FilterCadence) {
-                if (item.Action == FilterCadenceAction.Dither) {
-                    preDither = true;
-                    continue;
+            try {
+                bool preDither = false;
+                foreach (IFilterCadenceItem item in FilterCadence) {
+                    if (item.Action == FilterCadenceAction.Dither) {
+                        preDither = true;
+                        continue;
+                    }
+
+                    IExposure exposure = target.AllExposurePlans[item.ReferenceIdx];
+                    if (exposure.Rejected) { continue; }
+
+                    exposure.PreDither = preDither;
+                    FilterCadence.SetLastSelected(item);
+                    return exposure;
                 }
-
-                IExposure exposure = target.AllExposurePlans[item.ReferenceIdx];
-                if (exposure.Rejected) { continue; }
-
-                exposure.PreDither = preDither;
-                FilterCadence.SetLastSelected(item);
-                return exposure;
+            } catch (Exception ex) {
+                TSLogger.Warning($"exception in override exposure selector for target '{target.Name}' at time {atTime}: {ex.Message}, aborting target");
+                return null;
             }
 
-            // Fail safe ... should not happen
-            string msg = $"unexpected: no acceptable exposure plan in override exposure selector for target '{target.Name}' at time {atTime}";
-            TSLogger.Error(msg);
-            throw new Exception(msg);
+            TSLogger.Warning($"no acceptable exposure plan in override exposure selector for target '{target.Name}' at time {atTime}");
+            return null;
         }
 
         public void ExposureTaken(IExposure exposure) {